+++ src/config/config.ts
@@ -6,7 +6,7 @@ export default {
   hostVOT: "vot-api.toil.cc/v1",
   userAgent:
     "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/124.0.0.0 YaBrowser/24.6.0.0 Safari/537.36",
-  componentVersion: "24.6.4.580",
+  componentVersion: "24.7.0.2377",
   hmac: "bt8xH3VOlb4mqf0nqAibnDOoiPlXsisf",
   defaultDuration: 343,
 };