+++ changelog.md
@@ -1,3 +1,9 @@
+# 1.0.3
+
+- The behavior of the `data` field has been changed in case of a request error, now an error is written to it, if there is one
+- Some default types of any have been replaced with unknown
+- Removed extra console.log
+
 # 1.0.2
 
 - Added Reddit New & Reddit Old support